backend/requirements.txt
fastapi 
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic
pydantic-settings
python-dotenv
pandas
numpy
yfinance
scikit-learn